requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

hikari==2.0.0.dev120
hikari-lightbulb==2.3.3
hikari-miru==3.1.1
# Editable install with no version control (Pillow==9.0.1)
-e /usr/lib/python3/dist-packages/Pillow-9.0.1.egg-info
python-dotenv==1.0.0
# Editable install with no version control (requests==2.25.1)
-e /usr/lib/python3/dist-packages/requests-2.25.1.egg-info
uvloop==0.17.0
